Added requirements file, and updated program layout by seperating functions into their own files
