+++ requirements.txt
@@ -1,0 +1,7 @@
+
+#Dependencies
+selenium==4.9.0
+webdriver-manager==3.8.6
+beautifulsoup4==4.12.2
+requests==2.30.0
+flask==3.0.3
